% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_package.R
\name{change_Degree}
\alias{change_Degree}
\title{Change_Degree}
\usage{
change_Degree(longitude, lattitude, df)
}
\arguments{
\item{longitude}{the column in a dataframe of longitude values}

\item{df}{the original dataframe that will be manipulated}

\item{latitude}{the column in a dataframe of latitude values}
}
\value{
edited dataframe with longitude and latitude values changed
}
\description{
converts degree format from decimals degrees minutes to decimal degrees
}
\keyword{decimal}
\keyword{degrees}
\keyword{latitude}
\keyword{longitude}
